Started framing the form
This really just has the outline of the form, nothing really in it that can be said to be anywhere near an actual form.  Need to look closer at what I need to do for forms to get this to work properly.

I seem to have forgotten half of what I used to know about these things.  Hopefully all of it falls into place with a little bit of looking.