Creating basic structure
index.php - this will call the controller if the site is setup.
config.data.php - this holds the config data.
controller/index.php - this is the main controller.
model/index.php - this is a placeholder, probably will not be used.
view/index.php - this is a placeholder, and probably won't be used.

None of these files are in a functional state right now.
`config.data.php` might be the closest to what we want to have though.