Initial Structure Setup
I am trying to setup the initial structure.  This is starting to look at
a few things with the current setup, but isn't really doing anything
yet:

index.php:

This is the front end controller which calls `controller/index.php` when
`config.data.php` is properly setup.

config.data.php:

This file contains the configuration data.  Some sample data is
currently there, but it has not been tested yet.

Validation of this data will need to be done.  I think doing that would
probably be a `model` job.

controller/index.php:

This is the actual controller which gets called by `index.php` when
`config.data.php` is properly setup.

model/index.php:

This file is currently a placeholder file.  I think that if it would do
anything, it likely would be mostly to say "you're not supposed to be
here".

view/index.php

Placeholder file, it may well never be used.  And like the `model`
version probably would just say "go away, you're not supposed to be
here".