Getting Check Config Started
model/check-config.php:

New, it is supposed to check the configuration.  For some reason it
isn't working as expected.  Need to figure that out.  Will need to issue
it.

config.data.php:

Slight change to formatting, and trying to get variable name to be
correct.  Need to figure out why that's not working.

index.php:

Getting check-config properly called.