Fix scope issues, and diagnosis code
config.data.php:

Added diagnosis code to see if variable was getting set.  It was.

index.php:

Likewise with the diagnosis code here.  Variable is indeed set.

model/check-config.php:

Realised the reason for the variable not being found in the called
functions is due to the fact that the scope of the variable does not
fall into the functions.

Now we are passing the array as a paramater to the function.