move actions to separate file
Move the actions file to a separate file to make things easier to
maintain, so when action list changes, only one file needs to be
updated.

This may need some rethinking down the road, as we could be just listing
the actions as a list, or something like that, but this will do for now.