Setup database connection.
Right now this should error out.  I know that the connection will not
work the way that I have this setup.

If the error is that the connection fails, that should be fine.  If the
error is about syntax, I think I'll be able to fix it.

It is also possible that this is not getting called at all right now, so
if I get no errors, that probably means that it is not getting called.