Moving includes to subdirectory.
Trying to clean up the includes we call.

This is an intirm step to cleaning up the includes (thus the directory
structure) so that things work better.  There are files which are not
primary listed here.

We have copied to the includes directory, and edited the old versions so
that when they get called, they will display an error.

This will be a bit of a process, as I think I have missed a few files
that should go into the includes directory.