Renamed createdb.sql -> users.sql
I am currently looking at the create sql scripts as separate, though I
think in the end we will need to manage them in a reasonably "one step"
process.

This contains (at the time) user and profile information.  These are
going to end up getting separated in a later version.