Created docs for creating database
This is the file which is start of documentation of our process for
setting this up.

I have creation of database setup here.

This is likely not complete, but other than possibly having some of the
documentation not properly setup, I think this is enough of a start to
get this working now.