Added user verification table
This table will store verification data for verifying email (or maybe in
the future other methods) which will allow us to verify email addresses
in a secure way.

Currently no "method" is going to be listed in this table, as we are
only verifying email addresses.  I need to figure out how I am going to
handle failed verifications.