Added PHP Docblock.
I need to look at how to better handle this, but I am writing this for
now as it is because the docblock needs to be added.

There is something wrong with what I'm doing with creating the
databases, and it is only ending up running (from what I can see) one
(the latest) of the table creations.