Fixed to load multiple databases.
This (though not changes here) now is installing multiple databases.

The databases were only getting one created because of an error I had
created with a different file (servervars related), so this was not
producing the correct information for me.

This does appear to be creating the databases correctly now.