Added class to form for "checkboxLabel"
The form (now that it was loading the CSS) was displaying in a way that
ended up breaking the display.  The `:` is no longer displayed, and the
label is displayed as `inline` not `inline-block`.  Need to look to see
if the `16em` `inline-block` `width` is wide enough for the form for
submitting listings.