Clean up old code
We have removed old code which was mostly comments.

We fixed some issues which we hadn't seen, related to the creation of
the verification hash.  Though this probably didn't matter, it is
creating the hash much as we expect.

TODO: Fix array to string conversion to use impode() where possible.