Start of processing verification code
This is the start of the verification code.  I think that we will need
to handle this a bit better in the future as I'm starting to see that we
will run into issues.

Part of the problem is that each action hitting the cotroller currently
is calling a single file for managing the data received, which executes
the code itself, and then a file for displaying content.

The problem that I forsee coming up is that related handling routines
will be used for different actions.  Login/Register/Verify are probably
highly related issues.